+++ src/main/java/org/springframework/beans/factory/InitializingBean.java
@@ -1,0 +1,10 @@
+package org.springframework.beans.factory;
+
+/**
+ * @author derekyi
+ * @date 2020/11/29
+ */
+public interface InitializingBean {
+
+	void afterPropertiesSet() throws Exception;
+}