+++ src/main/java/org/springframework/beans/factory/config/BeanDefinition.java
@@ -14,6 +14,10 @@ public class BeanDefinition {
 
 	private PropertyValues propertyValues;
 
+	private String initMethodName;
+
+	private String destroyMethodName;
+
 	public BeanDefinition(Class beanClass) {
 		this(beanClass, null);
 	}
@@ -37,5 +41,21 @@ public class BeanDefinition {
 
 	public void setPropertyValues(PropertyValues propertyValues) {
 		this.propertyValues = propertyValues;
+	}
+
+	public String getInitMethodName() {
+		return initMethodName;
+	}
+
+	public void setInitMethodName(String initMethodName) {
+		this.initMethodName = initMethodName;
+	}
+
+	public String getDestroyMethodName() {
+		return destroyMethodName;
+	}
+
+	public void setDestroyMethodName(String destroyMethodName) {
+		this.destroyMethodName = destroyMethodName;
 	}
 }