+++ src/main/java/org/springframework/beans/factory/config/ConfigurableBeanFactory.java
@@ -12,4 +12,9 @@ public interface ConfigurableBeanFactory extends HierarchicalBeanFactory, Single
 	 * @param beanPostProcessor
 	 */
 	void addBeanPostProcessor(BeanPostProcessor beanPostProcessor);
+
+	/**
+	 * 销毁单例bean
+	 */
+	void destroySingletons();
 }