+++ src/main/java/org/springframework/beans/factory/support/DefaultSingletonBeanRegistry.java
@@ -1,9 +1,13 @@
 package org.springframework.beans.factory.support;
 
+import org.springframework.beans.BeansException;
+import org.springframework.beans.factory.DisposableBean;
 import org.springframework.beans.factory.config.SingletonBeanRegistry;
 
 import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.Map;
+import java.util.Set;
 
 /**
  * @author derekyi
@@ -13,6 +17,8 @@ public class DefaultSingletonBeanRegistry implements SingletonBeanRegistry {
 
 	private Map<String, Object> singletonObjects = new HashMap<>();
 
+	private final Map<String, DisposableBean> disposableBeans = new HashMap<>();
+
 	@Override
 	public Object getSingleton(String beanName) {
 		return singletonObjects.get(beanName);
@@ -20,5 +26,21 @@ public class DefaultSingletonBeanRegistry implements SingletonBeanRegistry {
 
 	protected void addSingleton(String beanName, Object singletonObject) {
 		singletonObjects.put(beanName, singletonObject);
+	}
+
+	public void registerDisposableBean(String beanName, DisposableBean bean) {
+		disposableBeans.put(beanName, bean);
+	}
+
+	public void destroySingletons() {
+		Set<String> beanNames = disposableBeans.keySet();
+		for (String beanName : beanNames) {
+			DisposableBean disposableBean = disposableBeans.remove(beanName);
+			try {
+				disposableBean.destroy();
+			} catch (Exception e) {
+				throw new BeansException("Destroy method on bean with name '" + beanName + "' threw an exception", e);
+			}
+		}
 	}
 }