+++ src/main/java/org/springframework/beans/factory/xml/XmlBeanDefinitionReader.java
@@ -32,6 +32,8 @@ public class XmlBeanDefinitionReader extends AbstractBeanDefinitionReader {
 	public static final String CLASS_ATTRIBUTE = "class";
 	public static final String VALUE_ATTRIBUTE = "value";
 	public static final String REF_ATTRIBUTE = "ref";
+	public static final String INIT_METHOD_ATTRIBUTE = "init-method";
+	public static final String DESTROY_METHOD_ATTRIBUTE = "destroy-method";
 
 	public XmlBeanDefinitionReader(BeanDefinitionRegistry registry) {
 		super(registry);
@@ -74,6 +76,8 @@ public class XmlBeanDefinitionReader extends AbstractBeanDefinitionReader {
 					String id = bean.getAttribute(ID_ATTRIBUTE);
 					String name = bean.getAttribute(NAME_ATTRIBUTE);
 					String className = bean.getAttribute(CLASS_ATTRIBUTE);
+					String initMethodName = bean.getAttribute(INIT_METHOD_ATTRIBUTE);
+					String destroyMethodName = bean.getAttribute(DESTROY_METHOD_ATTRIBUTE);
 
 					Class<?> clazz = null;
 					try {
@@ -89,6 +93,8 @@ public class XmlBeanDefinitionReader extends AbstractBeanDefinitionReader {
 					}
 
 					BeanDefinition beanDefinition = new BeanDefinition(clazz);
+					beanDefinition.setInitMethodName(initMethodName);
+					beanDefinition.setDestroyMethodName(destroyMethodName);
 
 					for (int j = 0; j < bean.getChildNodes().getLength(); j++) {
 						if (bean.getChildNodes().item(j) instanceof Element) {