+++ src/main/java/org/springframework/context/ConfigurableApplicationContext.java
@@ -14,4 +14,15 @@ public interface ConfigurableApplicationContext extends ApplicationContext {
 	 * @throws BeansException
 	 */
 	void refresh() throws BeansException;
+
+	/**
+	 * 关闭应用上下文
+	 */
+	void close();
+
+	/**
+	 * 向虚拟机中注册一个钩子方法，在虚拟机关闭之前执行关闭容器等操作
+	 */
+	void registerShutdownHook();
+
 }