+++ src/main/java/org/springframework/context/support/AbstractApplicationContext.java
@@ -83,5 +83,27 @@ public abstract class AbstractApplicationContext extends DefaultResourceLoader i
 	}
 
 	public abstract ConfigurableListableBeanFactory getBeanFactory();
+
+	public void close() {
+		doClose();
+	}
+
+	public void registerShutdownHook() {
+		Thread shutdownHook = new Thread() {
+			public void run() {
+				doClose();
+			}
+		};
+		Runtime.getRuntime().addShutdownHook(shutdownHook);
+
+	}
+
+	protected void doClose() {
+		destroyBeans();
+	}
+
+	protected void destroyBeans() {
+		getBeanFactory().destroySingletons();
+	}
 }