+++ src/test/java/org/springframework/test/ioc/InitAndDestoryMethodTest.java
@@ -1,0 +1,17 @@
+package org.springframework.test.ioc;
+
+import org.junit.Test;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
+
+/**
+ * @author derekyi
+ * @date 2020/11/29
+ */
+public class InitAndDestoryMethodTest {
+
+	@Test
+	public void testInitAndDestroyMethod() throws Exception {
+		ClassPathXmlApplicationContext applicationContext = new ClassPathXmlApplicationContext("classpath:init-and-destroy-method.xml");
+		applicationContext.registerShutdownHook();  //或者手动关闭 applicationContext.close();
+	}
+}