+++ src/test/java/org/springframework/test/ioc/bean/Person.java
@@ -1,16 +1,37 @@
 package org.springframework.test.ioc.bean;
 
+import org.springframework.beans.factory.DisposableBean;
+import org.springframework.beans.factory.InitializingBean;
+
 /**
  * @author derekyi
  * @date 2020/11/24
  */
-public class Person {
+public class Person implements InitializingBean, DisposableBean {
 
 	private String name;
 
 	private int age;
 
 	private Car car;
+
+	public void customInitMethod() {
+		System.out.println("I was born in the method named customInitMethod");
+	}
+
+	public void customDestroyMethod() {
+		System.out.println("I died in the method named customDestroyMethod");
+	}
+
+	@Override
+	public void afterPropertiesSet() throws Exception {
+		System.out.println("I was born in the method named afterPropertiesSet");
+	}
+
+	@Override
+	public void destroy() throws Exception {
+		System.out.println("I died in the method named destroy");
+	}
 
 	public String getName() {
 		return name;
@@ -44,4 +65,5 @@ public class Person {
 				", car=" + car +
 				'}';
 	}
+
 }